Refactor classes in `entities` package
