+++ app/src/main/java/com/hit/playpal/chatrooms/domain/utils/exceptions/MessagesNotFoundException.java
@@ -1,7 +1,7 @@
 package com.hit.playpal.chatrooms.domain.utils.exceptions;
 
-public class MessagesNotFound extends Exception{
-    public MessagesNotFound() {
+public class MessagesNotFoundException extends Exception{
+    public MessagesNotFoundException() {
         super("Messages not found in the chat room.");
     }
 }
