+++ app/src/main/java/com/hit/playpal/chatrooms/domain/utils/exceptions/RoomNotFoundException.java
@@ -1,7 +1,7 @@
 package com.hit.playpal.chatrooms.domain.utils.exceptions;
 
-public class RoomNotFound extends Exception {
-    public RoomNotFound() {
+public class RoomNotFoundException extends Exception {
+    public RoomNotFoundException() {
         super("Room not found");
     }
 }
