--- app/src/main/java/com/hit/playpal/entities/chats/enums/AppearancePolicy.java
@@ -1,6 +1,0 @@
-package com.hit.playpal.entities.chats.enums;
-
-public enum AppearancePolicy {
-    REGULARS,
-    MODERATORS,
-}
