--- app/src/main/java/com/hit/playpal/entities/chats/enums/JoiningPolicy.java
@@ -1,6 +1,0 @@
-package com.hit.playpal.entities.chats.enums;
-
-public enum JoiningPolicy {
-    REGULARS,
-    MODERATORS,
-}
