+++ app/src/main/java/com/hit/playpal/entities/users/UserPrivate.java
@@ -2,8 +2,10 @@ package com.hit.playpal.entities.users;
 
 import com.google.firebase.firestore.PropertyName;
 
+/**
+ * Entity class that represents a user's private information.
+ */
 public class UserPrivate {
-
     @PropertyName("email") private String mEmail;
     @PropertyName("email") public String getEmail() { return mEmail; }
     @PropertyName("email") public void setEmail(String iEmail) { mEmail = iEmail; }
